+++ src/commands/enter.sh
@@ -1,0 +1,3 @@
+enter_command () {
+    sh shell.sh
+}