+++ src/commands/stop.sh
@@ -1,0 +1,4 @@
+stop_command () {
+    docker-compose stop
+    docker-compose rm -f
+}