+++ src/commands/update.sh
@@ -1,0 +1,5 @@
+update_command () {
+    stop_command
+    git pull origin master
+    sh update-docker.sh
+}