+++ src/devilbox-cli.sh
@@ -1,0 +1,31 @@
+##########################
+# For test purposes only #
+##########################
+
+source config.sh
+
+source utils/exit.sh
+source utils/messages.sh
+source utils/filesystem.sh
+source utils/prompt.sh
+source utils/choices.sh
+source utils/config.sh
+source utils/path.sh
+
+source commands/config/common.sh
+source commands/config/apache.sh
+source commands/config/php.sh
+source commands/config/mysql.sh
+source commands/config/docroot.sh
+source commands/config/projects.sh
+source commands/config/databases.sh
+source commands/config.sh
+source commands/enter.sh
+source commands/help.sh
+source commands/open.sh
+source commands/run.sh
+source commands/stop.sh
+source commands/update.sh
+source commands/version.sh
+
+source main.sh