+++ CHANGELOG.md
@@ -4,10 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0] - 2019-03-08
+### Added
+    - Build script
+    - New commands: enter, config, open, run, stop, update
+### Changed
+    - Restructure the whole project with separate files
+    - Use printf instead of echo
+    - Versions helpers are now called choices helpers
+    - Possibility to change default devilbox path with DEVILBOX_PATH in profile source
+    - Fancy messages output
+
+## [0.0.3] - 2019-03-08
+### Changed
+    - Sed regex compatibility with characters
+
 ## [0.0.2] - 2019-03-08
 ### Changed
     - Use short function declarations
-    - format helpers with colors
+    - Format helpers with colors
 
 ## [0.0.1] - 2019-03-08
 ### Added